requirements.txt
langchain
langchain-groq
langchain-community
langchain_google_genai
streamlit
streamlit-lottie
groq
pydantic
python-dotenv
tiktoken
flask
flask-cors
flask-restful
selenium
webdriver-manager
beautifulsoup4
requests
plotly
google-generativeai>=0.3.0
google-search-results>=2.4.2
google-genai
google
serpapi
tiktoken